Improve code readability Aws node
